# openapi.yaml
openapi: 3.0.3
info:
  title: University Services API
  version: 1.0.0
  description: 'The central API platform for university operations.

    Covers Authentication, Students, Programs & Courses, Faculty, Library, AI
    Assistant, API Key Management, and UAE Government Reporting domains.


    ## Authentication

    Most endpoints require a Bearer JWT token obtained via `POST /auth/login`.

    Government reporting endpoints (`/gov/*`) require an `X-API-Key` header instead.


    Get a JWT token: `POST /auth/login` with your email and password.

    Use JWT: Add header `Authorization: Bearer <token>` to all requests.

    Use API Key: Add header `X-API-Key: uak_<your_key>` — service role only.


    ## Test Credentials (local / seeded environment)

    | Role | Email | Password |

    |------|-------|----------|

    | Admin | admin@university.ac.ae | Admin@1234 |

    | Student | student1@university.ac.ae | Student@1234 |

    | Faculty | faculty1@university.ac.ae | Faculty@1234 |

    | Librarian | librarian@university.ac.ae | Librarian@1234 |


    ## API-Led Connectivity Architecture

    Follows MuleSoft API-led connectivity:

    - **System Layer**: Raw data access (student records, course catalog, book inventory)

    - **Process Layer**: Business logic (enrollment validation, borrowing rules, capacity
    checks)

    - **Experience Layer**: Consumer-tailored responses (student portal, faculty dashboard)

...


    ## AI Library Assistant

...
    OpenAI.

...

    '
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via OAuth2 authorization flow
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
      description: Number of results per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: 'Field to sort by. Prefix with - for descending. Example: sort=-created_at'
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 25
        total_items:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The request body contains invalid fields
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
              example:
              - field: email
                issue: Must be a valid university email address
    StudentBase:
      type: object
      required:
      - first_name
      - last_name
      - email
      - date_of_birth
      - program_id
      - enrollment_year
      properties:
        first_name:
          type: string
          example: Nazar
        last_name:
          type: string
          example: Kamaal
        email:
          type: string
          format: email
          example: nazar.kamaal@university.ac.ae
        date_of_birth:
          type: string
          format: date
          example: '1997-05-15'
        program_id:
          type: string
          description: Reference to the Programs domain — not embedded, fetched separately
          example: prog_cs_101
        enrollment_year:
          type: integer
          example: 2026
        status:
          type: string
          enum:
          - active
          - inactive
          - graduated
          - suspended
          default: active
    Student:
      allOf:
      - $ref: '#/components/schemas/StudentBase'
      - type: object
        properties:
          id:
            type: string
            example: stu_10042
            readOnly: true
          created_at:
            type: string
            format: date-time
            readOnly: true
          updated_at:
            type: string
            format: date-time
            readOnly: true
    StudentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    CourseBase:
      type: object
      required:
      - code
      - title
      - program_id
      - credits
      - semester
      - faculty_id
      - max_capacity
      properties:
        code:
          type: string
          example: CS301
        title:
          type: string
          example: Machine Learning Fundamentals
        description:
          type: string
          example: Introduction to supervised and unsupervised learning techniques
        program_id:
          type: string
          example: prog_cs_101
        credits:
          type: integer
          example: 3
        semester:
          type: string
          example: fall_2026
        faculty_id:
          type: string
          description: Reference to the Faculty domain
          example: fac_2001
        max_capacity:
          type: integer
          example: 40
    Course:
      allOf:
      - $ref: '#/components/schemas/CourseBase'
      - type: object
        properties:
          id:
            type: string
            example: crs_5010
            readOnly: true
          current_enrollment:
            type: integer
            description: Server-computed derived field. Read-only — never sent by
              the client.
            readOnly: true
            example: 27
          created_at:
            type: string
            format: date-time
            readOnly: true
          updated_at:
            type: string
            format: date-time
            readOnly: true
    CourseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    EnrollmentRequest:
      type: object
      required:
      - student_id
      properties:
        student_id:
          type: string
          example: stu_10042
    Enrollment:
      type: object
      properties:
        id:
          type: string
          example: enr_80001
          readOnly: true
        course_id:
          type: string
          example: crs_5010
        student_id:
          type: string
          example: stu_10042
        enrolled_at:
          type: string
          format: date-time
          readOnly: true
        status:
          type: string
          enum:
          - active
          - dropped
          - completed
          default: active
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient permissions for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state (e.g. duplicate email, already
        enrolled)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Semantically invalid — passes validation but violates a business
        rule (e.g. course at capacity)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- BearerAuth: []
paths:
...
  /programs/{program_id}/courses:
    get:
      tags:
      - Programs
      summary: List all courses within a program
      description: Convenience relationship endpoint. Equivalent to GET /courses?program_id={program_id}
      parameters:
      - name: program_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: semester
        in: query
        schema:
          type: string
        example: fall_2026
      responses:
        '200':
          description: Paginated list of courses in this program
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getProgramCourses
  /courses:
    get:
      tags:
      - Courses
      summary: List all courses
      description: Returns courses across all programs. Filter by program_id to narrow
        down.
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/SortParam'
      - name: program_id
        in: query
        schema:
          type: string
      - name: semester
        in: query
        schema:
          type: string
      - name: faculty_id
        in: query
        schema:
          type: string
      - name: available
        in: query
        schema:
          type: boolean
        description: Filter to only courses with open enrollment spots
      responses:
        '200':
          description: Paginated list of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: listCourses
    post:
      tags:
      - Courses
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseBase'
      responses:
        '201':
          description: Course created successfully
          headers:
            Location:
              schema:
                type: string
                example: /courses/crs_5010
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      operationId: createCourse
  /courses/{course_id}:
    parameters:
    - name: course_id
      in: path
      required: true
      schema:
        type: string
      example: crs_5010
    get:
      tags:
      - Courses
      summary: Get a specific course
      responses:
        '200':
          description: Course record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getCourse
    put:
      tags:
      - Courses
      summary: Full update of a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseBase'
      responses:
        '200':
          description: Course updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: replaceCourse
    delete:
      tags:
      - Courses
      summary: Remove a course
      responses:
        '204':
          description: Course deleted
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deleteCourse
  /courses/{course_id}/students:
    get:
      tags:
      - Courses
      summary: List all students enrolled in a course
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of students in this course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getCourseStudents
  /courses/{course_id}/enrollments:
    post:
      tags:
      - Courses
      summary: Enroll a student in a course
      description: 'Business rules enforced here (Process API layer):

        - Student must exist

        - Course must exist and have available capacity

        - Student must not already be enrolled in this course

        - Prerequisite checks (future implementation)

        '
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
      responses:
        '201':
          description: Student enrolled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      operationId: enrollStudent
  /courses/{course_id}/enrollments/{enrollment_id}:
    delete:
      tags:
      - Courses
      summary: Remove an enrollment (drop a course)
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: string
      - name: enrollment_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Enrollment removed successfully
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: dropEnrollment
...
  /faculty/{faculty_id}/courses:
    get:
      tags:
      - Faculty
      summary: Get all courses taught by a faculty member
      parameters:
      - name: faculty_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: semester
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Paginated list of courses for this faculty member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getFacultyCourses
...